+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Changed
 -------
 * Auto-add 'output' prefix to ``field_name`` parameter for
   downloading files
+* Auto-wrapp ``list:*`` fields into list if they are not already
 
 
 ==================